Add migrations for a bunch of Mikkeller bars. Make abstract MikkellerBar scraper and make specific bar scraper inherit from it.
