+++ app/Scraper/HavenBarScraper.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Scraper;
+
+class HavenBarScraper extends MikkellerScraper
+{
+    protected $url = 'http://mikkeller.dk/location/haven-bar/';
+}
