+++ app/Scraper/MikkellerBarCopenhagenScraper.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Scraper;
+
+class MikkellerBarCopenhagenScraper extends MikkellerScraper
+{
+    protected $url = 'http://mikkeller.dk/location/mikkeller-bar-viktoriagade-copenhagen/';
+}
