+++ app/Scraper/MikropolisScraper.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Scraper;
+
+class MikropolisScraper extends MikkellerScraper
+{
+    protected $url = 'http://mikkeller.dk/location/mikropolis/';
+}
