+++ app/TaplistService.php
@@ -12,8 +12,11 @@ class TaplistService
      */
     public function updateTaplists()
     {
+        Bar::where('name', 'Haven Bar')->first()->updateTaplist();
+        /*
         Bar::all()->each(function($bar) {
             $bar->updateTaplist();
         });
+         */
     }
 }
